moviesOMDb/requirements.txt
coreapi==2.3.3
Django==2.2.12
django-filter==2.2.0
djangorestframework==3.11.0
drf-writable-nested==0.6.0
gunicorn==20.0.4